toolkit2: use int for maxEntriesLocalHeap
git-svn-id: http://svn.terracotta.org/svn/ehcache@5830 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache_toolkit2/terracotta/bootstrap/src/main/java/org/terracotta/modules/ehcache/ToolkitInstanceFactoryImpl.java
@@ -136,7 +136,7 @@ public class ToolkitInstanceFactoryImpl implements ToolkitInstanceFactory {
     if (ehcacheConfig.getPinningConfiguration() != null) {
       builder.pinningStore(getPinningStoreForConfiguration(ehcacheConfig));
     }
-    builder.maxCountLocalHeap(ehcacheConfig.getMaxEntriesLocalHeap());
+    builder.maxCountLocalHeap((int)ehcacheConfig.getMaxEntriesLocalHeap());
     builder.maxBytesLocalHeap(ehcacheConfig.getMaxBytesLocalHeap());
     builder.maxBytesLocalOffheap(ehcacheConfig.getMaxBytesLocalOffHeap());
     builder.offheapEnabled(ehcacheConfig.isOverflowToOffHeap());
